requirements.txt
undetected-chromedriver==3.5.5
selenium==4.22.0
APScheduler==3.10.4
pandas==2.2.2
pyyaml==6.0.1
python-dotenv==1.0.1

